third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/sin.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] sin float16 0D scalar
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got 0
[FAIL] sin float16 1D constant tensor
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got -0
[FAIL] sin float16 1D tensor
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got 0
[FAIL] sin float16 2D tensor
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got 0
[FAIL] sin float16 3D tensor
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got 0
[FAIL] sin float16 4D tensor
  assert_array_approx_equals: test sin float16 property 0, expected -0.935546875 +/- 0.001953125, expected -0.935546875 but got 0
Harness: the test ran to completion.
